+++ lib/sns.dart
@@ -1,9 +1,11 @@
 import 'dart:async';
-import 'package:meta/meta.dart';
-import 'src/request.dart';
-import 'src/credentials.dart';
+
 import 'package:http_client/http_client.dart' as http;
+import 'package:meta/meta.dart';
 import 'package:xml/xml.dart';
+
+import 'src/credentials.dart';
+import 'src/request.dart';
 
 typedef Future<String> RequestExecutor(Map<String, String> parameter);
 
@@ -13,10 +15,10 @@ class Sns {
   Sns({Credentials credentials, http.Client httpClient, String region})
       : _credentials = credentials,
         _httpClient = httpClient,
-        _region=region{
-    assert(this._credentials != null);
-    assert(this._httpClient != null);
-    assert(this._region != null);
+        _region = region {
+    assert(_credentials != null);
+    assert(_httpClient != null);
+    assert(_region != null);
   }
   final Credentials _credentials;
   final http.Client _httpClient;
@@ -24,121 +26,116 @@ class Sns {
 
   /// execute real request
   Future<String> _sendRequest(Map<String, String> parameters) async {
-    final endpoint = 'https://sns.${this._region}.amazonaws.com/';
-    AwsResponse response = await new AwsRequestBuilder(
+    final endpoint = 'https://sns.$_region.amazonaws.com/';
+    final response = await new AwsRequestBuilder(
       method: 'POST',
       baseUrl: endpoint,
       formParameters: parameters,
-      credentials: this._credentials,
-      httpClient: this._httpClient,
+      credentials: _credentials,
+      httpClient: _httpClient,
     ).sendRequest();
     final respString = await response.readAsString();
-    print('resp:${respString}');
     response.validateStatus();
     return respString;
   }
 
   /// return an Endpoint of arn
-  SnsEndpoint endpoint(String arn) => new SnsEndpoint(this._sendRequest, arn);
+  SnsEndpoint endpoint(String arn) => new SnsEndpoint(_sendRequest, arn);
 
   /// create an Endpoint with push token
   /// implements of https://docs.aws.amazon.com/sns/latest/api/API_CreatePlatformEndpoint.html
-  Future<SnsEndpoint> createEndpoint({
-    @required String applicationArn,
-    @required String pushToken,
-    String userData:''}) async {
-    assert(applicationArn!='');
-    assert(pushToken!='');
-    Map<String, String> parameters = {
+  Future<SnsEndpoint> createEndpoint(
+      {@required String applicationArn,
+      @required String pushToken,
+      String userData = ''}) async {
+    assert(applicationArn != '');
+    assert(pushToken != '');
+    final parameters = <String, String>{
       'Action': 'CreatePlatformEndpoint',
       'PlatformApplicationArn': applicationArn,
       'Token': pushToken,
       'CustomUserData': userData,
       'Version': '2010-03-31'
     };
-    XmlDocument xml = parse(await this._sendRequest(parameters));
+    final xml = parse(await _sendRequest(parameters));
     final endpointArn = xml.findAllElements('EndpointArn').first.text;
-    return this.endpoint(endpointArn);
+    return endpoint(endpointArn);
   }
 
-  SnsTopic topic(String arn) => new SnsTopic(this._sendRequest, arn);
+  SnsTopic topic(String arn) => new SnsTopic(_sendRequest, arn);
 
   /// Create a Topic
   /// implements of https://docs.aws.amazon.com/sns/latest/api/API_CreateTopic.html
   Future<SnsTopic> createTopic(String name) async {
-    Map<String, String> parameters = {
+    final parameters = <String, String>{
       'Action': 'CreateTopic',
       'Name': name,
       'Version': '2010-03-31'
     };
-    XmlDocument xml = parse(await this._sendRequest(parameters));
+    final xml = parse(await _sendRequest(parameters));
     final topicArn = xml.findAllElements('TopicArn').first.text;
-    return this.topic(topicArn);
+    return topic(topicArn);
   }
 
   /// Get subscription with arn
-  SnsSubscription subscription(String arn) => new SnsSubscription(this._sendRequest, arn);
-
+  SnsSubscription subscription(String arn) =>
+      new SnsSubscription(_sendRequest, arn);
 }
 
 /// The device endpoint
 class SnsEndpoint {
   /// A new endpoint of device
-  SnsEndpoint(
-      RequestExecutor sendRequest,
-      String arn) : _sendRequest = sendRequest,
-                    _arn = arn{
-    assert(sendRequest!=null);
-    assert(arn!=null);
+  SnsEndpoint(RequestExecutor sendRequest, String arn)
+      : _sendRequest = sendRequest,
+        _arn = arn {
+    assert(sendRequest != null);
+    assert(arn != null);
   }
   final RequestExecutor _sendRequest;
   final String _arn;
 
   /// The endpoint arg
-  String get arn => this._arn;
+  String get arn => _arn;
 
   /// push message to this endpoint
   /// implements of https://docs.aws.amazon.com/sns/latest/api/API_Publish.html
   /// return MessageId type:String
   Future<String> pushNotification(String body) async {
-    Map<String, String> parameters = {
+    final parameters = <String, String>{
       'Action': 'Publish',
-      'TargetArn': this._arn,
+      'TargetArn': _arn,
       'Message': body,
       'Version': '2010-03-31'
     };
-    XmlDocument xml = parse(await this._sendRequest(parameters));
+    final xml = parse(await this._sendRequest(parameters));
     final messageId = xml.findAllElements('MessageId').first.text;
     return messageId;
   }
-
 }
 
 /// SNS's Topic
 class SnsTopic {
   /// The SNS Topic
-  SnsTopic(
-      RequestExecutor sendRequest,
-      String arn) : _sendRequest = sendRequest,
-        _arn = arn{
-    assert(sendRequest!=null);
-    assert(arn!=null);
+  SnsTopic(RequestExecutor sendRequest, String arn)
+      : _sendRequest = sendRequest,
+        _arn = arn {
+    assert(sendRequest != null);
+    assert(arn != null);
   }
   final RequestExecutor _sendRequest;
   final String _arn;
-
 
   /// push message to this Topic
   /// implements of https://docs.aws.amazon.com/sns/latest/api/API_Publish.html
   /// return MessageId type:String
   Future<String> pushNotification(String body) async {
-    Map<String, String> parameters = {
+    final parameters = <String, String>{
       'Action': 'Publish',
-      'TargetArn': this._arn,
+      'TargetArn': _arn,
       'Message': body,
       'Version': '2010-03-31'
     };
-    XmlDocument xml = parse(await this._sendRequest(parameters));
+    final xml = parse(await this._sendRequest(parameters));
     final messageId = xml.findAllElements('MessageId').first.text;
     return messageId;
   }
@@ -147,33 +144,27 @@ class SnsTopic {
   /// implements of https://docs.aws.amazon.com/sns/latest/api/API_Subscribe.html
   /// return SubscriptionArn type:String
   Future<SnsSubscription> subscribe(String endpointArn) async {
-    Map<String, String> parameters = {
+    final parameters = <String, String>{
       'Action': 'Subscribe',
-      'TopicArn': this._arn,
+      'TopicArn': _arn,
       'Endpoint': endpointArn,
       'Protocol': 'application',
       'Version': '2010-03-31'
     };
-    XmlDocument xml = parse(await this._sendRequest(parameters));
+    final xml = parse(await this._sendRequest(parameters));
     final subscriptionArn = xml.findAllElements('SubscriptionArn').first.text;
-    return new SnsSubscription(
-        this._sendRequest,
-        subscriptionArn
-    );
+    return new SnsSubscription(_sendRequest, subscriptionArn);
   }
-
 }
 
 /// The SNS subscription
 class SnsSubscription {
   /// init subscription
-  SnsSubscription(
-      RequestExecutor sendRequest,
-      String arn
-      ): _sendRequest = sendRequest,
-        _arn = arn{
-    assert(sendRequest!=null);
-    assert(arn!=null);
+  SnsSubscription(RequestExecutor sendRequest, String arn)
+      : _sendRequest = sendRequest,
+        _arn = arn {
+    assert(sendRequest != null);
+    assert(arn != null);
   }
 
   final RequestExecutor _sendRequest;
@@ -181,12 +172,11 @@ class SnsSubscription {
 
   /// unsubcribe this subscription
   Future unsubscribe() async {
-    Map<String, String> parameters = {
+    final parameters = <String, String>{
       'Action': 'Unsubscribe',
-      'SubscriptionArn': this._arn,
+      'SubscriptionArn': _arn,
       'Version': '2010-03-31'
     };
     await this._sendRequest(parameters);
   }
-
 }